fixed copying pubkey

+++ postinstall.js
@@ -50,10 +50,16 @@ if (fs.existsSync(path.join(wellKnownDir, "route.ts"))) {
     appPath + "/node_modules/error-aware-client/api/.well-known/route.ts",
     path.join(wellKnownDir, "route.ts")
   );
-   fs.copyFileSync(
-     appPath + "/node_modules/error-aware-client/api/.well-known/public-key.ts",
-     path.join(wellKnownDir, "public-key.ts")
-   );
+}
+
+if (fs.existsSync(path.join(wellKnownDir, "public-key.ts"))) {
+  console.log("Public Key already exists. Skipping copy.");
+} else {
+  console.log(`Copying Public Key to ${wellKnownDir}`);
+  fs.copyFileSync(
+    appPath + "/node_modules/error-aware-client/api/.well-known/public-key.ts",
+    path.join(wellKnownDir, "public-key.ts")
+  );
 }
 
 console.log("Post-install script completed.");
